Allow Bookmarkable Tabs with Optional Query Parameters
Implement a bookmarking feature by enabling deep linking to tabs with
optional query parameters. A deep link is basically a link that directly
references a deeper structure and parameters of the app in the URL. It
is often bookmarked or shared to have a convenient entry point into the
app for a certain task or action. The selected tab should be reflected
in the URL but the tab can also be omitted, for example, when we
initially navigate to the resume page.